Rename config directory from cf to base - renamed import alias from cfconfig to config - updated path in kustomization.yaml - updated README.md
