+++ config/base/config.go
@@ -1,4 +1,4 @@
-package cf
+package base
 
 import (
 	"os"
