+++ controllers/workloads/integration/suite_integration_test.go
@@ -8,7 +8,7 @@ import (
 	"code.cloudfoundry.org/cf-k8s-controllers/controllers/workloads/fake"
 
 	workloadsv1alpha1 "code.cloudfoundry.org/cf-k8s-controllers/apis/workloads/v1alpha1"
-	cfconfig "code.cloudfoundry.org/cf-k8s-controllers/config/cf"
+	config "code.cloudfoundry.org/cf-k8s-controllers/config/base"
 	. "code.cloudfoundry.org/cf-k8s-controllers/controllers/workloads"
 
 	. "github.com/onsi/ginkgo"
@@ -80,9 +80,9 @@ var _ = BeforeSuite(func() {
 		Client: k8sManager.GetClient(),
 		Scheme: k8sManager.GetScheme(),
 		Log:    ctrl.Log.WithName("controllers").WithName("CFApp"),
-		ControllerConfig: &cfconfig.ControllerConfig{
+		ControllerConfig: &config.ControllerConfig{
 			KpackImageTag: "image/registry/tag",
-			CFProcessDefaults: cfconfig.CFProcessDefaults{
+			CFProcessDefaults: config.CFProcessDefaults{
 				MemoryMB:           500,
 				DefaultDiskQuotaMB: 512,
 			},
@@ -97,7 +97,7 @@ var _ = BeforeSuite(func() {
 		Client: k8sManager.GetClient(),
 		Scheme: k8sManager.GetScheme(),
 		Log:    ctrl.Log.WithName("controllers").WithName("CFBuild"),
-		ControllerConfig: &cfconfig.ControllerConfig{
+		ControllerConfig: &config.ControllerConfig{
 			KpackImageTag: "image/registry/tag",
 		},
 		RegistryAuthFetcher: NewRegistryAuthFetcher(registryAuthFetcherClient),
